fix migration conflicts
