+++ apps/backend-ops/src/firestoreMigrations/0027.ts
@@ -4,6 +4,7 @@ import { getStorageBucketName } from 'apps/backend-functions/src/internals/fireb
 import { Credit } from '@blockframes/utils/common-interfaces';
 import { get } from 'https';
 import { sanitizeFileName } from '@blockframes/utils/file-sanitizer';
+import { OldImgRef, OldPublicUser, OldPublicOrganization } from './old-types';
 
 const EMPTY_REF: OldImgRef = {
   ref: '',
@@ -187,36 +188,4 @@ const updateImgRef = async (
     console.log('Empty ref');
     return EMPTY_REF;
   }
-}
-
-interface OldImgRef {
-  ref: string;
-  urls: {
-    original: string;
-    fallback?: string;
-    xs?: string;
-    md?: string;
-    lg?: string;
-  };
-}
-
-interface OldPublicUser {
-  uid: string;
-  email: string;
-  avatar?: OldImgRef;
-  watermark?: OldImgRef;
-  firstName?: string;
-  lastName?: string;
-  orgId?: string;
-}
-
-interface OldPublicOrganization {
-  id: string;
-  denomination: OldDenomination;
-  logo: OldImgRef;
-}
-
-interface OldDenomination {
-    full: string;
-    public?: string;
 }
